+++ samples/Java/TestJavaClient/OrderDlg.java
@@ -435,7 +435,7 @@ public class OrderDlg extends JDialog {
 
             // set order fields
             m_order.action(m_action.getText());
-            m_order.totalQuantity(Integer.parseInt( m_totalQuantity.getText() ));
+            m_order.totalQuantity(Double.parseDouble( m_totalQuantity.getText() ));
             m_order.orderType(m_orderType.getText());
             m_order.lmtPrice(parseStringToMaxDouble( m_lmtPrice.getText()));
             m_order.auxPrice(parseStringToMaxDouble( m_auxPrice.getText()));
