+++ samples/Java/apidemo/Test.java
@@ -62,7 +62,7 @@ public class Test implements EWrapper {
 			double dividendsToLastTradeDate) {
 	}
 
-	@Override public void orderStatus(int orderId, String status, int filled, int remaining, double avgFillPrice, int permId, int parentId, double lastFillPrice, int clientId, String whyHeld) {
+	@Override public void orderStatus(int orderId, String status, double filled, double remaining, double avgFillPrice, int permId, int parentId, double lastFillPrice, int clientId, String whyHeld) {
 	}
 
 	@Override public void openOrder(int orderId, Contract contract, Order order, OrderState orderState) {
@@ -74,7 +74,7 @@ public class Test implements EWrapper {
 	@Override public void updateAccountValue(String key, String value, String currency, String accountName) {
 	}
 
-	@Override public void updatePortfolio(Contract contract, int position, double marketPrice, double marketValue, double averageCost, double unrealizedPNL, double realizedPNL, String accountName) {
+	@Override public void updatePortfolio(Contract contract, double position, double marketPrice, double marketValue, double averageCost, double unrealizedPNL, double realizedPNL, String accountName) {
 	}
 
 	@Override public void updateAccountTime(String timeStamp) {
@@ -146,7 +146,7 @@ public class Test implements EWrapper {
 	@Override public void commissionReport(CommissionReport commissionReport) {
 	}
 
-	@Override public void position(String account, Contract contract, int pos, double avgCost) {
+	@Override public void position(String account, Contract contract, double pos, double avgCost) {
 	}
 
 	@Override public void positionEnd() {
