+++ samples/Java/samples/rfq/SimpleWrapper.java
@@ -113,7 +113,7 @@ public class SimpleWrapper implements EWrapper {
 		logIn("tickEFP");
 	}
 
-	public void orderStatus(int orderId, String status, int filled, int remaining,
+	public void orderStatus(int orderId, String status, double filled, double remaining,
 			double avgFillPrice, int permId, int parentId, double lastFillPrice,
 			int clientId, String whyHeld) {
 		logIn("orderStatus");    	
@@ -131,7 +131,7 @@ public class SimpleWrapper implements EWrapper {
 		logIn("updateAccountValue");
 	}
 
-	public void updatePortfolio(Contract contract, int position, double marketPrice, double marketValue,
+	public void updatePortfolio(Contract contract, double position, double marketPrice, double marketValue,
 			double averageCost, double unrealizedPNL, double realizedPNL, String accountName) {
 		logIn("updatePortfolio");
 	}
@@ -233,7 +233,7 @@ public class SimpleWrapper implements EWrapper {
 	}
 
 	
-	public void position(String account, Contract contract, int pos, double avgCost) {
+	public void position(String account, Contract contract, double pos, double avgCost) {
 		logIn("position");
 	}
